@@ -1,16 +1,9 @@
 require File.expand_path('../boot', __FILE__)
 
-require 'active_record/railtie'
 require 'action_controller/railtie'
-require 'action_mailer/railtie'
 require 'sprockets/railtie'
 
-if defined?(Bundler)
-  # If you precompile assets before deploying to production, use this line
-  Bundler.require *Rails.groups(:assets => %w(development test))
-  # If you want your assets lazily compiled in production, use this line
-  # Bundler.setup(:default, :assets, Rails.env)
-end
+Bundler.require *Rails.groups(assets: %w(development test))
 
 module ScoutSublimevideo
   class Application < Rails::Application
@@ -21,16 +14,22 @@ module ScoutSublimevideo
     # http://ileitch.github.com/2012/03/24/rails-32-code-reloading-from-lib.html
     config.watchable_dirs['lib'] = [:rb]
 
+    # Enable the asset pipeline
+    config.assets.enabled = true
+
+    # Version of your assets, change this if you want to expire all your assets
+    config.assets.version = '1.0'
+
+    # http://guides.rubyonrails.org/asset_pipeline.html#precompiling-assets
+    # For faster asset precompiles, you can partially load your application
+    # by setting config.assets.initialize_on_precompile to false
+    # in config/application.rb, though in that case templates cannot see
+    # application objects or methods. Heroku requires this to be false.
+    config.assets.initialize_on_precompile = false
+
     # Only load the plugins named here, in the order given (default is alphabetical).
     # :all can be used as a placeholder for all plugins not explicitly named.
     # config.plugins = [ :exception_notification, :ssl_requirement, :all ]
-
-    # Activate observers that should always be running.
-    # config.active_record.observers = :cacher, :garbage_collector, :forum_observer
-
-    # Set Time.zone default to the specified zone and make Active Record auto-convert to this zone.
-    # Run "rake -D time" for a list of tasks for finding time zone names. Default is UTC.
-    # config.time_zone = 'Central Time (US & Canada)'
 
     # The default locale is :en and all translations from config/locales/*.rb,yml are auto loaded.
     # config.i18n.load_path += Dir[Rails.root.join('my', 'locales', '*.{rb,yml}').to_s]
@@ -44,24 +43,6 @@ module ScoutSublimevideo
 
     # Enable escaping HTML in JSON.
     config.active_support.escape_html_entities_in_json = true
-
-    # Use SQL instead of Active Record's schema dumper when creating the database.
-    # This is necessary if your schema can't be completely dumped by the schema dumper,
-    # like if you have constraints or database-specific column types
-    # config.active_record.schema_format = :sql
-
-    # Enforce whitelist mode for mass assignment.
-    # This will create an empty whitelist of attributes available for mass-assignment for all models
-    # in your app. As such, your models will need to explicitly whitelist or blacklist accessible
-    # parameters by using an attr_accessible or attr_protected declaration.
-    config.active_record.whitelist_attributes = true
-
-    # Enable the asset pipeline
-    config.assets.enabled = true
-    config.assets.initialize_on_precompile = false
-    # Version of your assets, change this if you want to expire all your assets
-    config.assets.version = '1.0'
-    config.assets.js_compressor = :closure
   end
 end
 
