@@ -1,7 +1,8 @@
 RSpec.configure do |config|
   config.treat_symbols_as_metadata_keys_with_true_values = true
   config.run_all_when_everything_filtered = true
-  config.filter_run_including focus: true
+  config.filter_run_including focus: ENV['FOCUS'] != 'false'
   config.mock_with :rspec
-  config.fail_fast = true
+  config.fail_fast = ENV['FAST_FAIL'] != 'false'
+  config.order = ENV['ORDER'] || 'random'
 end
