@@ -1,12 +1,26 @@
-begin
-  # use `bundle install --standalone' to get this...
-  require_relative '../bundle/bundler/setup'
-rescue LoadError
-  # fall back to regular bundler if the developer hasn't bundled standalone
-  require 'bundler'
-  Bundler.setup
+$LOAD_PATH.unshift("#{Dir.pwd}/app") unless $LOAD_PATH.include?("#{Dir.pwd}/app")
+Dir['app/**/'].each do |dir|
+  path = "#{Dir.pwd}/#{dir}"
+  $LOAD_PATH.unshift(path) unless $LOAD_PATH.include?(path)
 end
 
+ENV["RAILS_ENV"] ||= 'test'
+
+unless defined?(Rails)
+  module Rails
+    def self.root; Pathname.new(File.expand_path('')); end
+    def self.env; 'test'; end
+  end
+end
+
+unless defined?(Librato)
+  module Librato
+    def self.method_missing(*args)
+      true
+    end
+  end
+end
+
+require 'bundler/setup'
 require_relative 'config/rspec'
 require_relative 'support/stubs'
-require_relative 'support/fixtures_helpers'
