@@ -1,7 +1,10 @@
 require 'fast_spec_helper'
+require 'rails/railtie'
 require 'carrierwave'
-require File.expand_path('spec/config/carrierwave')
-require File.expand_path('app/uploaders/screenshot_uploader')
+require 'config/carrierwave'
+require 'support/fixtures_helpers'
+
+require 'uploaders/screenshot_uploader'
 
 describe ScreenshotUploader do
   let(:screenshot) { stub(id: '0123456789', site: stub(t: 'site_token'), u: 'http://sublimevideo.net') }
@@ -9,13 +12,9 @@ describe ScreenshotUploader do
   let(:uploader)   { described_class.new(screenshot, :f) }
 
   before do
-    stub_rails
     uploader.store!(image)
   end
-  after do
-    uploader.remove!
-    Dir.delete(Rails.root.join('screenshots')) if File.directory?(Rails.root.join('screenshots'))
-  end
+  after { uploader.remove! }
 
   it "is stored in ENV['S3_SCREENSHOTS_BUCKET'] bucket" do
     uploader.fog_directory.should eq ENV['S3_SCREENSHOTS_BUCKET']
@@ -29,8 +28,8 @@ describe ScreenshotUploader do
     uploader.file.content_type.should eq 'image/jpeg'
   end
 
-  it "begins with a timestamp" do
-    uploader.file.filename.should =~ /\A\d{10}\-/
+  it "begins with the model id" do
+    uploader.file.filename.should =~ /\A0123456789\-/
   end
 
   it "contains the URL parameterized" do
@@ -42,13 +41,6 @@ describe ScreenshotUploader do
   end
 
   describe 'carousel version' do
-    before do
-      described_class.enable_processing = true
-    end
-    after do
-      described_class.enable_processing = false
-    end
-
     describe 'the carousel version' do
       it "resize to fill the image to 1100x825" do
         uploader.carousel.should have_dimensions(1100, 825)
